requirements/base.txt
# Bleeding edge Django
Django==2.2.2
channels==2.2.0

# Configuration
django-environ==0.4.5
whitenoise==4.1.2
brotlipy==0.7.0

# Models
django-model-utils==3.1.1

# Rest apis
djangorestframework==3.9.4
django-filter==1.1.0
drf_ujson==1.2.0

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.4

# Time zones support
pytz==2019.1

celery==4.3.0
# Pin kombu version until issue https://github.com/celery/kombu/issues/870 is fixed
kombu==4.5.0
# we use msgpack since it outperforms yaml and json de/serializers
msgpack_python==0.5.6

# Utils
django-extensions==2.1.7
django-watchman==0.16.0
django-debug-toolbar==1.11

# Admin
django-admin-list-filter-dropdown==1.0.2

# extra apps
django-comments-xtd==2.4.0
django-contrib-comments==1.9.1
django-simple-history==2.7.2
django-taggit==1.1.0
django-tagulous==0.14.0

channels-redis==2.4.0
hiredis==1.0.0
msgpack==0.6.1
